refactor(sched): one 64-bit per-task protection-key rights word
The per-task field was named and sized for x86's PKRU. aarch64's POR_EL0
is the same concept with a wider register (64 bits of 4-bit fields versus
32 bits of 2-bit fields), and the policy around both is identical:
snapshot on switch-out, load on switch-in, inherit across fork, reset at
exec.

Naming and sizing it for one arch would have forced either a second field
or a rename once the aarch64 lane lands, and a second field is two owners
of one policy. Widened to u64 and renamed before the first arch ships, so
neither happens.

+++ crates/kernel/sched/src/pkey_rights/hw_none.rs
@@ -5,8 +5,8 @@
 /// # C: O(1)
 pub fn supported() -> bool { false }
 /// # C: O(1)
-pub fn init_value() -> u32 { 0 }
+pub fn init_value() -> u64 { 0 }
 /// # C: O(1)
-pub fn read_live() -> u32 { 0 }
+pub fn read_live() -> u64 { 0 }
 /// # C: O(1)
-pub fn write_live(v: u32) { let _ = v; }
+pub fn write_live(v: u64) { let _ = v; }
